requirements.txt
numpy
torch>=1.1
torchvision>=0.4
scikit-learn
yacs
tqdm
matplotlib
flask
Pillow
urllib3
gunicorn